fix: adjust portal conversion
