+++ screen/testdata/mock_server.c
@@ -13,6 +13,7 @@
 #include "../../wlr-screencopy-unstable-v1-client-protocol.h"
 #include "../../linux-dmabuf-unstable-v1-server-protocol.h"
 
+#define ZWLR_SCREENCOPY_FRAME_V1_BUFFER 0
 #define ZWLR_SCREENCOPY_FRAME_V1_LINUX_DMABUF 5
 #define ZWLR_SCREENCOPY_FRAME_V1_BUFFER_DONE 6
 #define ZWLR_SCREENCOPY_FRAME_V1_READY 3
@@ -21,6 +22,7 @@
 static struct wl_display *mock_display;
 static dev_t mock_dev;
 static uint64_t mock_modifier;
+static int use_shm;
 
 struct zwlr_screencopy_frame_v1_interface {
     void (*copy)(struct wl_client *, struct wl_resource *, struct wl_resource *);
@@ -50,7 +52,11 @@ static const struct zwlr_screencopy_frame_v1_interface frame_impl = {
 static void handle_capture_output(struct wl_client *client, struct wl_resource *resource, uint32_t id, struct wl_resource *output) {
     struct wl_resource *frame = wl_resource_create(client, &zwlr_screencopy_frame_v1_interface, 3, id);
     wl_resource_set_implementation(frame, &frame_impl, NULL, NULL);
-    wl_resource_post_event(frame, ZWLR_SCREENCOPY_FRAME_V1_LINUX_DMABUF, WL_SHM_FORMAT_ARGB8888, 64, 64);
+    if (use_shm) {
+        wl_resource_post_event(frame, ZWLR_SCREENCOPY_FRAME_V1_BUFFER, WL_SHM_FORMAT_ARGB8888, 64, 64, 256);
+    } else {
+        wl_resource_post_event(frame, ZWLR_SCREENCOPY_FRAME_V1_LINUX_DMABUF, WL_SHM_FORMAT_ARGB8888, 64, 64);
+    }
     wl_resource_post_event(frame, ZWLR_SCREENCOPY_FRAME_V1_BUFFER_DONE);
 }
 
@@ -142,13 +148,55 @@ static void bind_output(struct wl_client *client, void *data, uint32_t version, 
     wl_resource_set_implementation(res, NULL, NULL, NULL);
 }
 
+static void shm_pool_destroy(struct wl_client *client, struct wl_resource *resource) {
+    wl_resource_destroy(resource);
+}
+
+static void shm_pool_create_buffer(struct wl_client *client, struct wl_resource *resource, uint32_t id, int32_t offset, int32_t width, int32_t height, int32_t stride, uint32_t format) {
+    struct wl_resource *buf = wl_resource_create(client, &wl_buffer_interface, 1, id);
+    wl_resource_set_implementation(buf, NULL, NULL, NULL);
+}
+
+static void shm_pool_resize(struct wl_client *client, struct wl_resource *resource, int32_t size) { }
+
+static const struct wl_shm_pool_interface shm_pool_impl = {
+    .destroy = shm_pool_destroy,
+    .create_buffer = shm_pool_create_buffer,
+    .resize = shm_pool_resize,
+};
+
+static void shm_destroy(struct wl_client *client, struct wl_resource *resource) {
+    wl_resource_destroy(resource);
+}
+
+static void shm_create_pool(struct wl_client *client, struct wl_resource *resource, uint32_t id, int32_t fd, int32_t size) {
+    close(fd);
+    struct wl_resource *pool = wl_resource_create(client, &wl_shm_pool_interface, 1, id);
+    wl_resource_set_implementation(pool, &shm_pool_impl, NULL, NULL);
+}
+
+static const struct wl_shm_interface shm_impl = {
+    .destroy = shm_destroy,
+    .create_pool = shm_create_pool,
+};
+
+static void bind_shm(struct wl_client *client, void *data, uint32_t version, uint32_t id) {
+    struct wl_resource *res = wl_resource_create(client, &wl_shm_interface, 1, id);
+    wl_resource_set_implementation(res, &shm_impl, NULL, NULL);
+}
+
 void run_mock_server(const char *socket, uint32_t maj, uint32_t min, uint64_t modifier) {
     mock_dev = makedev(maj, min);
     mock_modifier = modifier;
+    use_shm = (maj == 0 && min == 0 && modifier == 0);
     mock_display = wl_display_create();
     wl_display_add_socket(mock_display, socket);
     wl_global_create(mock_display, &wl_output_interface, 2, NULL, bind_output);
-    wl_global_create(mock_display, &zwp_linux_dmabuf_v1_interface, 4, NULL, bind_dmabuf);
+    if (!use_shm) {
+        wl_global_create(mock_display, &zwp_linux_dmabuf_v1_interface, 4, NULL, bind_dmabuf);
+    } else {
+        wl_global_create(mock_display, &wl_shm_interface, 1, NULL, bind_shm);
+    }
     wl_global_create(mock_display, &zwlr_screencopy_manager_v1_interface, 3, NULL, bind_screencopy_manager);
     wl_display_run(mock_display);
     wl_display_destroy(mock_display);
